Remove obsolete ops files to secure rep admin API
Diego v2.0.0 removed the `admin_api.require_tls` property from the
rep and rep_windows jobs, making these operations files obsolete.

Note that the removal of these files would be a breaking change to the
filesystem interface for cf-deployment and so should be incorporated
only when ready to cut v2.0.0 of cf-deployment.